requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.1
Requests==2.32.5
selenium==4.39.0
